requirements.txt
# Slack ChatBot Dependencies
slack-sdk>=3.25.0
slack-bolt>=1.18.0
python-dotenv>=1.0.0
flask>=2.3.0
requests>=2.31.0

# LLM Integration - DeepSeek (OpenAI-compatible)
openai>=1.0.0

# Confluence Integration
atlassian-python-api>=3.41.0
beautifulsoup4>=4.12.0
lxml>=4.9.0

# Optional: For enhanced features
colorama>=0.4.4
click>=8.0.0
rich>=10.0.0